# backend/openapi.json
{
  "components": {
    "schemas": {
      "Action": {
        "description": "What the agent did, or what ground truth says it should have done.\n\nESCALATE is only ever an *agent* action. Ground truth is always APPROVE or REJECT,\nbecause every synthetic invoice carries a deterministic correct answer.",
        "enum": [
          "APPROVE",
          "REJECT",
          "ESCALATE"
        ],
        "title": "Action",
        "type": "string"
      },
...
      "AuditLogEntryOut": {
        "description": "One hash-chained row. `prev_hash`/`hash` make tampering with history\ndetectable \u2014 see docs/lanes/vp.md: \"sha256(prev_hash +\ncanonical_json(payload))\". Read-only; nothing in this API ever mutates\nan existing row.",
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "prev_hash": {
            "title": "Prev Hash",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ts",
          "actor",
          "actor_type",
          "event_type",
          "entity_type",
          "entity_id",
          "payload",
          "prev_hash",
          "hash"
        ],
        "title": "AuditLogEntryOut",
        "type": "object"
      },
      "AuditLogPage": {
        "description": "`GET /api/v1/audit-log`'s response: the usual pagination envelope,\nplus a chain-verification result computed fresh on every call \u2014 this is\nwhat makes tamper-evidence demonstrable on screen rather than claimed in\na docstring (docs/lanes/vp.md).\n\n`chain_verified_scope` says exactly what `chain_valid` covers: `\"full\"`\nwhen every row in `audit_log` was recomputed from `GENESIS_HASH`,\n`\"page\"` if the table ever grows large enough that a full recompute on\nevery request stops being cheap and this endpoint falls back to\nverifying only the returned page \u2014 never silently verifying less than\nit claims.",
        "properties": {
          "chain_valid": {
            "title": "Chain Valid",
            "type": "boolean"
          },
          "chain_verified_scope": {
            "enum": [
              "full",
              "page"
            ],
            "title": "Chain Verified Scope",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "chain_valid",
          "chain_verified_scope"
        ],
        "title": "AuditLogPage",
        "type": "object"
      },
      "AuditSampleOut": {
        "description": "Mirrors `shared.contracts.AuditSample` field-for-field.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "reviewer_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          },
          "sampled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sampled At"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewVerdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "sample_id",
          "decision_id",
          "agent_id",
          "sampled_at",
          "reviewed_at",
          "reviewer",
          "verdict",
          "reviewer_action"
        ],
        "title": "AuditSampleOut",
        "type": "object"
      },
      "AuditSampleReview": {
        "description": "Request body for `POST /api/v1/audit-samples/{sample_id}/review`.\n\nOnly ADMIN or REVIEWER may call this (`app/deps.py`) \u2014 reviewing sampled\ndecisions is REVIEWER's actual job (ADR-0009); AUDITOR stays read-only.",
        "properties": {
          "reason": {
            "description": "What the reviewer found, and why",
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "reviewer_action": {
            "$ref": "#/components/schemas/Action"
          },
          "verdict": {
            "$ref": "#/components/schemas/ReviewVerdict"
          }
        },
        "required": [
          "verdict",
          "reviewer_action",
          "reason"
        ],
        "title": "AuditSampleReview",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Every non-2xx response returns this shape as its JSON body.\n\n`code` is a short machine-readable slug (e.g. \"not_found\", \"forbidden\",\n\"validation_error\") for client branching; `message` is one sentence a human\nreads; `detail` carries whatever extra structure the specific failure needs\n(field errors, the resource id that was missing, etc.) and may be null.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AuditSampleOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditSampleOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Items per page, as requested",
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching records, independent of page size",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[AuditSampleOut]",
        "type": "object"
      },
...
      "ReviewVerdict": {
        "description": "A human reviewer's verdict on a sampled decision, from comparing the agent's\naction against what the reviewer determines was correct.",
        "enum": [
          "AGREED",
          "DISAGREED",
          "INCONCLUSIVE"
        ],
        "title": "ReviewVerdict",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The integration point for all four lanes: the Policy Engine, persistence and audit, the HTTP contract, and the hard ceiling. See docs/SYSTEM-EXPLAINED.md for the full design.",
    "title": "Adaptive AI Governance Platform \u2014 Backend",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit-log": {
      "get": {
        "description": "The complete hash-chained event log, newest first. Read-only \u2014\nnothing in this API ever mutates an existing row (that's the point).\n\nRecomputes the whole chain from `GENESIS_HASH` on every call \u2014\n`audit_log` is small enough in this system for that to be cheap \u2014 and\nreports the result as `chain_valid`/`chain_verified_scope` rather than\njust asserting immutability in a docstring. If the table ever grows\nlarge enough that a full recompute stops being cheap, this falls back\nto verifying only the returned page and says so via\n`chain_verified_scope`, instead of silently verifying less than it\nclaims.",
        "operationId": "list_audit_log_api_v1_audit_log_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Log",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit-samples": {
      "get": {
        "description": "List sampled decisions pulled for human review, newest first.\n\nOnce implemented: `SELECT ... ORDER BY sampled_at DESC`, with a\n`?pending=true` filter for the review queue view \u2014 samples are pulled\nat `sampling_rate_of(agent.current_rung)` (shared/constants.py) as\ndecisions are recorded, not on a schedule.",
        "operationId": "list_audit_samples_api_v1_audit_samples_get",
        "parameters": [
          {
            "description": "1-indexed page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditSampleOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Samples",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit-samples/{sample_id}/review": {
      "post": {
        "description": "Record a human review of one sampled decision. REVIEWER or ADMIN only.\n\nOnce implemented: writes `reviewed_at`/`reviewer`/`verdict`/\n`reviewer_action` onto the `audit_samples` row, and \u2014 if `verdict` is\n`DISAGREED` \u2014 emits `SAMPLE_REVIEW_DISAGREEMENT`\n(shared/reason_codes.py) into the agent's next trust evaluation, since\na disagreeing sample is itself evidence. This stub validates the sample\nexists and is still pending, then returns a copy reflecting the\nreview \u2014 it does not persist it or feed the trust engine.",
        "operationId": "review_audit_sample_api_v1_audit_samples__sample_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditSampleReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditSampleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The current role may not perform this action."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No resource with that id."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The resource is not in a state this action applies to."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Audit Sample",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}